requirements.txt
alembic==1.0.0
asn1crypto==0.24.0
astroid==1.6.4
atomicwrites==1.1.5
attrs==18.1.0
bcrypt==3.1.4
blinker==1.4
cffi==1.13.2
click==6.7
colorama==0.3.9
coverage==4.5.1
cryptography==2.8
factory-boy==3.2.1
Faker==13.0.0
Flask==1.0.2
Flask-Bcrypt==0.7.1
Flask-Cors==3.0.4
Flask-JWT-Extended==3.12.1
Flask-Login==0.4.1
Flask-Mail==0.9.1
flask-marshmallow==0.9.0
Flask-SQLAlchemy==2.3.2
Flask-WTF==0.14.2
gunicorn==20.1.0
idna==2.6
isort==4.3.4
itsdangerous==0.24
Jinja2==2.10.3
lazy-object-proxy==1.3.1
Mako==1.0.7
MarkupSafe==1.1.1
marshmallow==2.15.4
marshmallow-sqlalchemy==0.14.1
mccabe==0.6.1
more-itertools==4.3.0
pathlib2==2.3.2
pluggy==0.7.1
py==1.5.4
pycparser==2.19
PyJWT==1.6.4
pylint==1.9.1
pytest==3.7.1
python-dateutil==2.7.3
python-editor==1.0.3
six==1.13.0
SQLAlchemy==1.3.11
Werkzeug==0.16.0
wrapt==1.10.11
WTForms==2.1
